fix(build): close #341
issue
- rollup doesn't correctly handle mix modules (import and require) files which
  results in inline require files not getting included in final build.
- log4js uses dynamic requires

fix
- replace all require with import in code
- remove log4js dependency. (use same Logger as used in flowLSP code)

other changes
- add eslint rules to prevent this from happening in future.
