+++ .eslintignore
@@ -1,5 +1,4 @@
 flow-libs
-lib/flowNonLSP
 build
 rollup.config.js
 babel.config.js
