+++ lib/flowLSP/utils/importFresh.js
@@ -13,5 +13,7 @@ export default function importFresh(moduleId: string) {
       delete require.cache[filePath];
     }
   });
+  // NOTE: safe to use dynamic require here
+  // eslint-disable-next-line global-require, import/no-dynamic-require
   return require(moduleId);
 }
