+++ lib/flowNonLSP/.eslintrc.js
@@ -1,23 +1,15 @@
 module.exports = {
   root: true,
 
-  plugins: ['eslint-plugin-playlyfe', 'eslint-plugin-import'],
+  parser: 'babel-eslint',
 
-  extends: [
-    'plugin:playlyfe/js',
-    'plugin:playlyfe/flowtype',
-    'plugin:playlyfe/prettier',
-    'plugin:playlyfe/testing:jest',
-  ],
+  plugins: ['eslint-plugin-import'],
 
   env: {
     node: true,
   },
 
   rules: {
-    'no-negated-condition': 'off',
-    'no-undefined': 'off',
-
     /* Rules to prevent use of require in code: with rollup it's not safe to use */
     'global-require': 'error',
     'import/no-commonjs': 'error',
