+++ lib/flowNonLSP/flowLogging.js
@@ -1,12 +1,8 @@
-/**
- * Adds a gloabl that is used inside consoleAppender.js to output console messages
- * to the user, instead of to the developer console.
- */
+/* @flow */
 import * as vscode from 'vscode';
+import getOutputChannel from './utils/getOutputChannel';
 
-export function setupLogging(context): void {
-  const channel = vscode.window.createOutputChannel('Flow');
-  vscode.commands.registerCommand('flow.show-output', () => channel.show());
-  global.flowOutputChannel = channel;
+export function setupLogging(context: vscode.ExtensionContext): void {
+  const channel = getOutputChannel();
   context.subscriptions.push(channel);
 }
