+++ lib/flowNonLSP/pkg/flow-base/lib/FlowHelpers.js
@@ -20,7 +20,7 @@ import invariant from 'assert';
 import path from 'path';
 import os from 'os';
 
-import {getLogger} from '../../nuclide-logging/lib/main';
+import getLogger from '../../../utils/getLogger';
 const logger = getLogger();
 
 const flowConfigDirCache: LRUCache<string, Promise<?string>> = LRU({
