+++ lib/flowNonLSP/pkg/flow-base/lib/FlowProcess.js
@@ -16,7 +16,7 @@ import os from 'os';
 
 import {BehaviorSubject, Observable} from 'rxjs';
 
-import {getLogger} from '../../nuclide-logging/lib/main';
+import getLogger from '../../../utils/getLogger';
 const logger = getLogger();
 
 // import {track} from '../../nuclide-analytics';
