+++ lib/flowNonLSP/pkg/flow-base/lib/FlowRoot.js
@@ -21,7 +21,7 @@ import type {
 import {filter} from 'fuzzaldrin';
 import semver from 'semver';
 
-import {getLogger} from '../../nuclide-logging/lib/main';
+import getLogger from '../../../utils/getLogger';
 const logger = getLogger();
 
 import {
