--- lib/flowNonLSP/pkg/nuclide-logging/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "nuclide-logging",
-  "repository": "https://github.com/facebook/nuclide",
-  "main": "./lib/main.js",
-  "version": "0.0.0",
-  "description": "Provides logging on both Nuclide client and Nuclide server",
-  "nuclide": {
-    "packageType": "Node",
-    "testRunner": "npm"
-  },
-  "scripts": {
-    "test": "node ../nuclide-jasmine/bin/jasmine-node-transpiled spec"
-  }
-}
