+++ lib/flowNonLSP/utils/getLogger.js
@@ -1,0 +1,13 @@
+/* @flow */
+import getOutputChannel from './getOutputChannel';
+import Logger from '../../common/Logger';
+import { getLogLevel } from './util';
+
+export default function getLogger() {
+  console.log(getLogLevel());
+  return new Logger(
+    '',
+    getOutputChannel(),
+    getLogLevel(),
+  );
+}
