+++ lib/flowNonLSP/utils/util.js
@@ -39,6 +39,10 @@ export function getFileExtensions():Array<string> {
 	return workspace.getConfiguration('flow').get('fileExtensions')
 }
 
+export function getLogLevel() {
+	return workspace.getConfiguration('flow').get('logLevel')
+}
+
 export function shouldStopFlowOnExit():boolean {
 	return workspace.getConfiguration('flow').get('stopFlowOnExit')
 }
