+++ package.json
@@ -218,7 +218,6 @@
     "fs-plus": "^2.8.2",
     "fuzzaldrin": "^2.1.0",
     "lodash.debounce": "^4.0.8",
-    "log4js": "^0.6.37",
     "lru-cache": "^4.0.1",
     "mkdirp": "^0.5.1",
     "prettier": "1.18.2",
@@ -239,6 +238,7 @@
     "babel-jest": "24.8.0",
     "builtin-modules": "3.1.0",
     "eslint": "5.13.0",
+    "eslint-plugin-import": "2.17.3",
     "eslint-plugin-playlyfe": "6.2.1",
     "jest": "24.8.0",
     "rollup": "1.14.5",
